requirements.txt
streamlit==1.36.0
requests==2.32.4
feedparser==6.0.11
beautifulsoup4==4.13.4
lxml==6.0.0
pandas==2.3.1
altair==5.5.0
pyvis==0.3.2
transformers==4.41.2
tokenizers==0.19.1
torch==2.5.1
regex==2025.7.34
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
